Convert pattern-library entrypoint to TS
